pyproject.toml
[tool.poetry]
name = "aims-saas-adapter-kims-py"
version = "0.1.0"
description = ""
authors = [
    "Sandeep Wali <sandeep.wali@solumesl.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
azure-storage-blob = "^12.19.1"
chardet = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
